# packages/api-spec/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Manhaj API",
    "version": "1.0.0",
    "description": "Offline-First Student Question Bank API"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "authToken": {
        "type": "apiKey",
        "in": "header",
        "name": "x-auth-token"
      }
    },
    "schemas": {
      "StructuredSubject": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "moduleId": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "lectures": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Lecture"
            }
          }
        }
      },
      "StructuredModule": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "termId": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "subjects": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StructuredSubject"
            }
          }
        }
      },
      "StructuredTerm": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "gradeId": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "modules": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StructuredModule"
            }
          }
        }
      },
      "StructuredGrade": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "terms": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StructuredTerm"
            }
          }
        }
      },
      "Grade": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "Term": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "gradeId": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "Module": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "termId": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "Subject": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "moduleId": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "Lecture": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "subjectId": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "LectureDetails": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "subjectId": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "subject": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "format": "uuid"
              },
              "name": {
                "type": "string"
              }
            }
          },
          "lectureVideos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LectureVideo"
            }
          },
          "lectureFiles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LectureFile"
            }
          },
          "questions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Question"
            }
          }
        }
      },
      "Question": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "lectureId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "createdBy": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "questionText": {
            "type": "string"
          },
          "explanation": {
            "type": "string"
          },
          "source": {
            "type": "string",
            "enum": [
              "telegram_auto",
              "admin_manual"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "Choice": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "questionId": {
            "type": "string",
            "format": "uuid"
          },
          "choiceText": {
            "type": "string"
          },
          "isCorrect": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "LectureVideo": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "lectureId": {
            "type": "string",
            "format": "uuid"
          },
          "sourceName": {
            "type": "string"
          },
          "url": {
            "type": "string"
          },
          "duration": {
            "type": "integer"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "LectureFile": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "lectureId": {
            "type": "string",
            "format": "uuid"
          },
          "sourceName": {
            "type": "string"
          },
          "fileUrl": {
            "type": "string"
          },
          "fileType": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "VideoProgress": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "lectureVideoId": {
            "type": "string",
            "format": "uuid"
          },
          "positionSeconds": {
            "type": "integer"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/content/sync": {
      "get": {
        "summary": "Sync content for offline use",
        "tags": [
          "Content"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "authToken": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "since",
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "description": "Cursor for delta sync (ISO timestamp)"
          }
        ],
        "responses": {
          "200": {
            "description": "Content sync successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "grades": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Grade"
                      }
                    },
                    "terms": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Term"
                      }
                    },
                    "modules": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Module"
                      }
                    },
                    "subjects": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Subject"
                      }
                    },
                    "lectures": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Lecture"
                      }
                    },
                    "questions": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Question"
                      }
                    },
                    "choices": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Choice"
                      }
                    },
                    "lectureVideos": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/LectureVideo"
                      }
                    },
                    "lectureFiles": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/LectureFile"
                      }
                    },
                    "nextCursor": {
                      "type": "string",
                      "format": "date-time"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/content/hierarchy": {
      "get": {
        "summary": "Get complete nested content hierarchy",
        "tags": [
          "Content"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "authToken": []
          }
        ],
        "responses": {
          "200": {
            "description": "Structured content hierarchy retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StructuredGrade"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/content/lectures/{id}": {
      "get": {
        "summary": "Get lecture details with videos, files, and questions",
        "tags": [
          "Content"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "authToken": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lecture details retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LectureDetails"
                }
              }
            }
          },
          "404": {
            "description": "Lecture not found"
          }
        }
      }
    },
    "/content/lectures/{id}/videos": {
      "get": {
        "summary": "Get videos for a lecture",
        "tags": [
          "Content"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "authToken": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lecture videos retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LectureVideo"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/content/video-progress": {
      "post": {
        "summary": "Update video watch progress",
        "tags": [
          "Content"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "authToken": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "lectureVideoId",
                  "positionSeconds"
                ],
                "properties": {
                  "lectureVideoId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "positionSeconds": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Video progress updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VideoProgress"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}